Renaming gromet_memgraph.rs to database.rs
